+++ scripts/test/TemplateGenerator/Replace.t
@@ -12,12 +12,12 @@ use utf8;
 
 use vars (qw($Self));
 
-my $ConfigObject       = $Kernel::OM->Get('Kernel::Config');
 my $DynamicFieldObject = $Kernel::OM->Get('Kernel::System::DynamicField');
 my $BackendObject      = $Kernel::OM->Get('Kernel::System::DynamicField::Backend');
 my $TicketObject       = $Kernel::OM->Get('Kernel::System::Ticket');
+my $UserObject         = $Kernel::OM->Get('Kernel::System::User');
 
-# get helper object
+# Get helper object.
 $Kernel::OM->ObjectParamAdd(
     'Kernel::System::UnitTest::Helper' => {
         RestoreDatabase  => 1,
@@ -26,7 +26,9 @@ $Kernel::OM->ObjectParamAdd(
 );
 my $Helper = $Kernel::OM->Get('Kernel::System::UnitTest::Helper');
 
-my $RandomID = $Helper->GetRandomID();
+my $RandomID             = $Helper->GetRandomID();
+my $NotificationLanguage = 'en';
+my $UserLanguage         = 'de';
 
 my @DynamicFieldsToAdd = (
     {
@@ -77,7 +79,7 @@ for my $DynamicField (@DynamicFieldsToAdd) {
         'DynamicFieldAdd()',
     );
 
-    # remember added DynamicFields
+    # Remember added DynamicFields.
     $AddedDynamicFieldIds{$DynamicFieldID} = $DynamicField->{Name};
 
     my $DynamicFieldConfig = $DynamicFieldObject->DynamicFieldGet(
@@ -89,61 +91,77 @@ for my $DynamicField (@DynamicFieldsToAdd) {
         'DynamicFieldConfig must be a hash reference',
     );
 
-    # remember the DF config
+    # Remember the DF config.
     $DynamicFieldConfigs{ $DynamicField->{FieldType} } = $DynamicFieldConfig;
 }
 
-# create template generator after the dynamic field are created as it gathers all DF in the
-# constructor
+# Create template generator after the dynamic field are created as it gathers all DF in the
+# constructor.
 my $TemplateGeneratorObject = $Kernel::OM->Get('Kernel::System::TemplateGenerator');
 
 my $TestCustomerLogin = $Helper->TestCustomerUserCreate(
-    Language => 'en',
+    Language => $UserLanguage,
 );
 
-my $TestUserLogin = $Helper->TestUserCreate(
-    Language => 'en',
+my %TestCustomerData = $Kernel::OM->Get('Kernel::System::CustomerUser')->CustomerUserDataGet(
+    User => $TestCustomerLogin,
 );
 
-my %TestUser = $Kernel::OM->Get('Kernel::System::User')->GetUserData(
-    User => $TestUserLogin,
-);
+my @TestUsers;
+for ( 1 .. 4 ) {
+    my $TestUserLogin = $Helper->TestUserCreate(
+        Language => $UserLanguage,
+    );
+    my %TestUser = $UserObject->GetUserData(
+        User => $TestUserLogin,
+    );
+    push @TestUsers, \%TestUser;
+}
 
-my $TestUser2Login = $Helper->TestUserCreate(
-    Language => 'en',
-);
+# Create time for time tags check.
+my $SystemTime = $Kernel::OM->Create(
+    'Kernel::System::DateTime',
+    ObjectParams => {
+        String => '2017-07-05 11:00:00',
+    },
+)->ToEpoch();
 
-my %TestUser2 = $Kernel::OM->Get('Kernel::System::User')->GetUserData(
-    User => $TestUserLogin,
-);
+# Set the fixed time.
+$Helper->FixedTimeSet($SystemTime);
 
-my $TestUser3Login = $Helper->TestUserCreate(
-    Language => 'en',
+# Create test queue with escalation times.
+my $QueueID = $Kernel::OM->Get('Kernel::System::Queue')->QueueAdd(
+    Name                => 'Queue' . $RandomID,
+    ValidID             => 1,
+    GroupID             => 1,
+    FirstResponseTime   => 30,
+    FirstResponseNotify => 80,
+    UpdateTime          => 40,
+    UpdateNotify        => 80,
+    SolutionTime        => 50,
+    SolutionNotify      => 80,
+    SystemAddressID     => 1,
+    SalutationID        => 1,
+    SignatureID         => 1,
+    UserID              => 1,
+    Comment             => "Test Queue",
 );
-
-my %TestUser3 = $Kernel::OM->Get('Kernel::System::User')->GetUserData(
-    User => $TestUserLogin,
-);
-
-my $TestUser4Login = $Helper->TestUserCreate(
-    Language => 'en',
-);
-
-my %TestUser4 = $Kernel::OM->Get('Kernel::System::User')->GetUserData(
-    User => $TestUserLogin,
+$Self->True(
+    $QueueID,
+    "QueueID $QueueID - created"
 );
 
 my $TicketID = $TicketObject->TicketCreate(
     Title         => 'Some Ticket_Title',
-    Queue         => 'Raw',
+    QueueID       => $QueueID,
     Lock          => 'unlock',
     Priority      => '3 normal',
-    State         => 'closed successful',
+    State         => 'open',
     CustomerNo    => '123465',
     CustomerUser  => $TestCustomerLogin,
-    OwnerID       => $TestUser{UserID},
-    ResponsibleID => $TestUser2{UserID},
-    UserID        => $TestUser3{UserID},
+    OwnerID       => $TestUsers[0]->{UserID},
+    ResponsibleID => $TestUsers[1]->{UserID},
+    UserID        => $TestUsers[2]->{UserID},
 );
 $Self->IsNot(
     $TicketID,
@@ -177,6 +195,9 @@ my $ArticleBackendObject = $Kernel::OM->Get('Kernel::System::Ticket::Article')->
     ChannelName => 'Internal',
 );
 
+# Add 5 minutes for escalation times evaluation.
+$Helper->FixedTimeAddSeconds(300);
+
 my $ArticleID = $ArticleBackendObject->ArticleCreate(
     TicketID             => $TicketID,
     IsVisibleForCustomer => 0,
@@ -196,6 +217,10 @@ $Self->IsNot(
     undef,
     'ArticleCreate() ArticleID',
 );
+
+# Renew object because of transaction.
+$Kernel::OM->ObjectsDiscard( Objects => ['Kernel::System::Ticket'] );
+$TicketObject = $Kernel::OM->Get('Kernel::System::Ticket');
 
 my @Tests = (
     {
@@ -262,7 +287,7 @@ my @Tests = (
         },
         RichText => 0,
         Template => 'Test <OTRS_RESPONSIBLE_UserFirstname> <OTRS_RESPONSIBLE_nonexisting>',
-        Result   => "Test $TestUser2{UserFirstname} -",
+        Result   => "Test $TestUsers[1]->{UserFirstname} -",
     },
     {
         Name => 'OTRS_TICKET_RESPONSIBLE firstname',              # <OTRS_RESPONSIBLE_UserFirstname>
@@ -271,7 +296,7 @@ my @Tests = (
         },
         RichText => 0,
         Template => 'Test <OTRS_TICKET_RESPONSIBLE_UserFirstname> <OTRS_TICKET_RESPONSIBLE_nonexisting>',
-        Result   => "Test $TestUser2{UserFirstname} -",
+        Result   => "Test $TestUsers[1]->{UserFirstname} -",
     },
     {
         Name => 'OTRS owner firstname',                           # <OTRS_OWNER_*>
@@ -280,7 +305,7 @@ my @Tests = (
         },
         RichText => 0,
         Template => 'Test <OTRS_OWNER_UserFirstname> <OTRS_OWNER_nonexisting>',
-        Result   => "Test $TestUser{UserFirstname} -",
+        Result   => "Test $TestUsers[0]->{UserFirstname} -",
     },
     {
         Name => 'OTRS_TICKET_OWNER firstname',                    # <OTRS_OWNER_*>
@@ -289,7 +314,7 @@ my @Tests = (
         },
         RichText => 0,
         Template => 'Test <OTRS_TICKET_OWNER_UserFirstname> <OTRS_TICKET_OWNER_nonexisting>',
-        Result   => "Test $TestUser{UserFirstname} -",
+        Result   => "Test $TestUsers[0]->{UserFirstname} -",
     },
     {
         Name => 'OTRS current firstname',                         # <OTRS_CURRENT_*>
@@ -298,7 +323,7 @@ my @Tests = (
         },
         RichText => 0,
         Template => 'Test <OTRS_CURRENT_UserFirstname> <OTRS_CURRENT_nonexisting>',
-        Result   => "Test $TestUser3{UserFirstname} -",
+        Result   => "Test $TestUsers[2]->{UserFirstname} -",
     },
     {
         Name => 'OTRS ticket ticketid',                           # <OTRS_TICKET_*>
@@ -554,7 +579,84 @@ Line7</div>',
         Data     => {},
         RichText => 0,
         Template => 'Test <OTRS_NOTIFICATION_RECIPIENT_UserFullname> <OTRS_NOTIFICATION_RECIPIENT_nonexisting>',
-        Result   => "Test $TestUser4{UserFullname} -",
+        Result   => "Test $TestUsers[3]->{UserFullname} -",
+    },
+    {
+        Name     => 'OTRS <OTRS_TICKET_EscalationResponseTime>',
+        Data     => {},
+        RichText => 0,
+        Template => 'Test <OTRS_TICKET_EscalationResponseTime>',
+        Result   => "Test 07/05/2017 11:30",
+    },
+    {
+        Name     => 'OTRS <OTRS_TICKET_EscalationUpdateTime>',
+        Data     => {},
+        RichText => 0,
+        Template => 'Test <OTRS_TICKET_EscalationUpdateTime>',
+        Result   => "Test 07/05/2017 11:45",
+    },
+    {
+        Name     => 'OTRS <OTRS_TICKET_EscalationSolutionTime>',
+        Data     => {},
+        RichText => 0,
+        Template => 'Test <OTRS_TICKET_EscalationSolutionTime>',
+        Result   => "Test 07/05/2017 11:50",
+    },
+    {
+        Name     => 'OTRS <OTRS_TICKET_EscalationTimeWorkingTime>',
+        Data     => {},
+        RichText => 0,
+        Template => 'Test <OTRS_TICKET_EscalationTimeWorkingTime>',
+        Result   => "Test 25 m",
+    },
+    {
+        Name     => 'OTRS <OTRS_TICKET_EscalationTime>',
+        Data     => {},
+        RichText => 0,
+        Template => 'Test <OTRS_TICKET_EscalationTime>',
+        Result   => "Test 25 m",
+    },
+    {
+        Name     => 'OTRS <OTRS_TICKET_FirstResponseTimeWorkingTime>',
+        Data     => {},
+        RichText => 0,
+        Template => 'Test <OTRS_TICKET_FirstResponseTimeWorkingTime>',
+        Result   => "Test 25 m",
+    },
+    {
+        Name     => 'OTRS <OTRS_TICKET_FirstResponseTime>',
+        Data     => {},
+        RichText => 0,
+        Template => 'Test <OTRS_TICKET_FirstResponseTime>',
+        Result   => "Test 25 m",
+    },
+    {
+        Name     => 'OTRS <OTRS_TICKET_UpdateTimeWorkingTime>',
+        Data     => {},
+        RichText => 0,
+        Template => 'Test <OTRS_TICKET_UpdateTimeWorkingTime>',
+        Result   => "Test 40 m",
+    },
+    {
+        Name     => 'OTRS <OTRS_TICKET_UpdateTime>',
+        Data     => {},
+        RichText => 0,
+        Template => 'Test <OTRS_TICKET_UpdateTime>',
+        Result   => "Test 40 m",
+    },
+    {
+        Name     => 'OTRS <OTRS_TICKET_SolutionTimeWorkingTime>',
+        Data     => {},
+        RichText => 0,
+        Template => 'Test <OTRS_TICKET_SolutionTimeWorkingTime>',
+        Result   => "Test 45 m",
+    },
+    {
+        Name     => 'OTRS <OTRS_TICKET_SolutionTime>',
+        Data     => {},
+        RichText => 0,
+        Template => 'Test <OTRS_TICKET_SolutionTime>',
+        Result   => "Test 45 m",
     },
 );
 
@@ -570,8 +672,71 @@ for my $Test (@Tests) {
         DataAgent   => $Test->{DataAgent},
         RichText    => $Test->{RichText},
         TicketData  => \%Ticket,
-        UserID      => $TestUser3{UserID},
-        RecipientID => $TestUser4{UserID},
+        UserID      => $TestUsers[2]->{UserID},
+        RecipientID => $TestUsers[3]->{UserID},
+        Language    => $NotificationLanguage,
+    );
+    $Self->Is(
+        $Result,
+        $Test->{Result},
+        "$Test->{Name} - _Replace()",
+    );
+}
+
+# Set state to 'pending reminder'.
+$Success = $TicketObject->TicketStateSet(
+    State    => 'pending reminder',
+    TicketID => $TicketID,
+    UserID   => $TestUsers[2]->{UserID},
+);
+$Self->True(
+    $Success,
+    "TicketID $TicketID - set state to pending reminder successfully",
+);
+
+$Success = $TicketObject->TicketPendingTimeSet(
+    String   => '2017-07-06 10:00:00',
+    TicketID => $TicketID,
+    UserID   => 1,
+);
+$Self->True(
+    $Success,
+    "Set pending time successfully",
+);
+
+# Check 'UntilTime' and 'RealTillTimeNotUsed' tags (see bug#8301).
+@Tests = (
+    {
+        Name     => 'OTRS <OTRS_TICKET_UntilTime>',
+        Data     => {},
+        RichText => 0,
+        Template => 'Test <OTRS_TICKET_UntilTime>',
+        Result   => "Test 22 h 55 m",
+    },
+    {
+        Name     => 'OTRS <OTRS_TICKET_RealTillTimeNotUsed>',
+        Data     => {},
+        RichText => 0,
+        Template => 'Test <OTRS_TICKET_RealTillTimeNotUsed>',
+        Result   => "Test 07/06/2017 10:00",
+    }
+);
+
+%Ticket = $TicketObject->TicketGet(
+    TicketID      => $TicketID,
+    DynamicFields => 1,
+);
+
+for my $Test (@Tests) {
+    my $Result = $TemplateGeneratorObject->_Replace(
+        Text        => $Test->{Template},
+        Data        => $Test->{Data},
+        DataAgent   => $Test->{DataAgent},
+        RichText    => $Test->{RichText},
+        TicketData  => \%Ticket,
+        UserID      => $TestUsers[2]->{UserID},
+        RecipientID => $TestUsers[3]->{UserID},
+        Language    => $NotificationLanguage,
     );
     $Self->Is(
         $Result,